Update incorrect safeApprove to deprecated
